requirements.txt
requests>=2.32.0
fastapi>=0.115.0
uvicorn>=0.30.0
websocket-client>=1.8.0
